アラート共通化
