+++ RandomChoiceApp/Controller/EditViewController.swift
@@ -149,12 +149,4 @@ extension EditViewController {
             dismiss(animated: true, completion: nil)
         }
     }
-    
-    private func showAlertAllNilTextField() {
-        let alert = UIAlertController(title: AlertTitleLiteral.allTextEmpty, message: AlertMessageLiteral
-            .allTextEmpty, preferredStyle: .alert)
-        let cancelAction = UIAlertAction(title: AlertButtonLiteral.OK, style: .cancel, handler: nil)
-        alert.addAction(cancelAction)
-        present(alert, animated: true, completion: nil)
-    }
 }
